requirements.txt
streamlit
ollama
pypdf
requests
beautifulsoup4
chromadb
langchain-text-splitters
google-generativeai
